+++ src/app/compare/reducer.ts
@@ -3,7 +3,7 @@ import { DimItem } from 'app/inventory/item-types';
 import { showNotification } from 'app/notifications/notifications';
 import { getSelectionTree } from 'app/organizer/ItemTypeSelector';
 import { quoteFilterString } from 'app/search/query-parser';
-import { getInterestingSocketMetadatas, isD1Item } from 'app/utils/item-utils';
+import { getSpecialtySocketMetadata, isD1Item } from 'app/utils/item-utils';
 import { getArmorArchetype } from 'app/utils/socket-utils';
 import { ItemCategoryHashes, PlugCategoryHashes } from 'data/d2/generated-enums';
 import { ActionType, Reducer, getType } from 'typesafe-actions';
@@ -175,11 +175,13 @@ function initialCompareQuery(item: DimItem) {
       factors.push(`exactperk:${quoteFilterString(intrinsicName)}`);
     }
 
-    const modSlotMetadata = getInterestingSocketMetadatas(item);
+    let modSlotMetadata = getSpecialtySocketMetadata(item);
+    if (modSlotMetadata?.slotTag === 'artifice') {
+      modSlotMetadata = undefined;
+    }
+
     if (modSlotMetadata) {
-      for (const m of modSlotMetadata) {
-        factors.push(`modslot:${m.slotTag}`);
-      }
+      factors.push(`modslot:${modSlotMetadata.slotTag}`);
     }
 
     const archetype = getArmorArchetype(item);