+++ docs/CHANGELOG.md
@@ -1,7 +1,17 @@
 ## Next
 
+* Reduced the height of the automatic Max Light loadout on the Loadouts page.
+* Fix Loadout names being forced to uppercase.
+* Add supplied set bonuses to loadouts on the Loadouts page.
+* Fixed "Wrong Stat Minimums" showing up on loadouts without stat constraints.
+
+## 8.91.0 <span class="changelog-date">(2025-09-21)</span>
+
 * Updated item tiles to more closely match in-game tiles, with higher quality images.
 * Remove armor energy capacity from the item tile. It's just not that interesting these days.
+* Added Max power loadout to Loadouts tab
+* Items in loadout optimizer sets will now reflect changes like lock/unlock or being masterworked in-game.
+* hide artifact power since artifact power has been removed
 
 ## 8.90.1 <span class="changelog-date">(2025-09-18)</span>